requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
pytest==8.3.2